+++ src/test/javascript/class-scanner_test.ts
@@ -88,6 +88,12 @@ suite('Class', () => {
             if (p.type != null) {
               param.type = p.type;
             }
+            if (p.defaultValue != null) {
+              param.defaultValue = p.defaultValue;
+            }
+            if (p.rest != null) {
+              param.rest = p.rest;
+            }
             return param;
           });
         }
@@ -231,11 +237,34 @@ suite('Class', () => {
                 {
                   name: 'b',
                   type: '...Number',
+                  rest: true,
                   description: 'The second argument.',
                 }
               ],
               return: {
                 desc: 'The number 9, always.',
+                type: 'Number',
+              },
+            },
+            {
+              name: 'customInstanceFunctionWithParamDefault',
+              description: 'This is the description for ' +
+                  'customInstanceFunctionWithParamDefault.',
+              params: [
+                {
+                  name: 'a',
+                  type: 'Number',
+                  description: 'The first argument.',
+                },
+                {
+                  name: 'b',
+                  type: 'Number',
+                  defaultValue: '0',
+                  description: 'The second argument.',
+                }
+              ],
+              return: {
+                desc: 'The number 10, always.',
                 type: 'Number',
               },
             },
@@ -422,11 +451,34 @@ suite('Class', () => {
                 {
                   name: 'b',
                   type: '...Number',
+                  rest: true,
                   description: 'The second argument.',
                 }
               ],
               return: {
                 desc: 'The number 9, always.',
+                type: 'Number',
+              },
+            },
+            {
+              name: 'customInstanceFunctionWithParamDefault',
+              description: 'This is the description for ' +
+                  'customInstanceFunctionWithParamDefault.',
+              params: [
+                {
+                  name: 'a',
+                  type: 'Number',
+                  description: 'The first argument.',
+                },
+                {
+                  name: 'b',
+                  type: 'Number',
+                  defaultValue: '0',
+                  description: 'The second argument.',
+                }
+              ],
+              return: {
+                desc: 'The number 10, always.',
                 type: 'Number',
               },
             },