+++ src/test/polymer/polymer2-element-scanner-old-jsdoc_test.ts
@@ -416,13 +416,15 @@ namespaced name.`,
               {
                 name: 'customInstanceFunction',
                 description: '',
-                params: [], return: undefined
+                params: [],
+                return: undefined
               },
               {
                 name: 'customInstanceFunctionWithJSDoc',
                 description: 'This is the description for ' +
                     'customInstanceFunctionWithJSDoc.',
-                params: [], return: {
+                params: [],
+                return: {
                   desc: 'The number 5, always.',
                   type: 'Number',
                 },
@@ -431,9 +433,27 @@ namespaced name.`,
                 name: 'customInstanceFunctionWithParams',
                 description: '',
                 params: [
-                  {name: 'a', type: undefined, description: undefined},
-                  {name: 'b', type: undefined, description: undefined},
-                  {name: 'c', type: undefined, description: undefined}
+                  {
+                    name: 'a',
+                    type: undefined,
+                    defaultValue: undefined,
+                    rest: undefined,
+                    description: undefined
+                  },
+                  {
+                    name: 'b',
+                    type: undefined,
+                    defaultValue: undefined,
+                    rest: undefined,
+                    description: undefined
+                  },
+                  {
+                    name: 'c',
+                    type: undefined,
+                    defaultValue: undefined,
+                    rest: undefined,
+                    description: undefined
+                  }
                 ],
                 return: undefined,
               },
@@ -445,12 +465,22 @@ namespaced name.`,
                   {
                     name: 'a',
                     type: 'Number',
+                    defaultValue: undefined,
+                    rest: undefined,
                     description: 'The first argument',
                   },
-                  {name: 'b', type: 'Number', description: undefined},
+                  {
+                    name: 'b',
+                    type: 'Number',
+                    defaultValue: undefined,
+                    rest: undefined,
+                    description: undefined
+                  },
                   {
                     name: 'c',
                     type: 'Number',
+                    defaultValue: undefined,
+                    rest: undefined,
                     description: 'The third argument',
                   }
                 ],
@@ -463,7 +493,8 @@ namespaced name.`,
                 name: 'customInstanceFunctionWithParamsAndPrivateJSDoc',
                 description: 'This is the description for\n' +
                     'customInstanceFunctionWithParamsAndPrivateJSDoc.',
-                params: [], return: undefined,
+                params: [],
+                return: undefined,
               },
             ],
             warningUnderlines: [],